rename: S3Properties properties 패키지로 이동
