+++ src/main/java/sixgaezzang/sidepeek/config/S3Config.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import sixgaezzang.sidepeek.media.util.S3Properties;
+import sixgaezzang.sidepeek.config.properties.S3Properties;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
